# public/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MarketMate API",
    "version": "2.1.0",
    "description": "The MarketMate API provides real-time and historical market intelligence for cryptocurrency and forex markets. Access market regime analysis, trading signals, liquidity maps, and historical data to power your trading applications, dashboards, and automated strategies.\n\n## Authentication\nAll API endpoints require a valid API key passed via the `Authorization` header as a Bearer token. You can generate and manage API keys from your MarketMate dashboard.\n\n## Rate Limits\n- **Free tier**: 60 requests/minute\n- **Pro tier**: 300 requests/minute\n- **Enterprise tier**: 1,000 requests/minute\n\nRate limit headers are included in every response: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`.\n\n## WebSocket Support\nReal-time data is also available via WebSocket at `wss://marketmate-website.onrender.com/ws`. See the event schemas documentation for WebSocket event formats.\n\n## SDKs & Tools\nOfficial SDKs are available for Python, JavaScript/TypeScript, and Go. Postman and Insomnia collections can be downloaded from the developers portal.",
    "contact": {
      "name": "MarketMate API Support",
      "url": "https://marketmate-website.onrender.com/developers",
      "email": "api@marketmate.io"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://marketmate-website.onrender.com/terms"
    }
  },
  "servers": [
    {
...
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/liquidity": {
      "get": {
        "operationId": "getLiquidity",
        "summary": "Get liquidity analysis",
        "description": "Returns the current liquidity map for a given symbol, including liquidity pools (demand/supply levels with strength and volume) and recent liquidity sweeps. This data identifies where institutional orders are likely clustered and where stops have been triggered.",
        "tags": ["Market Data"],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "description": "Trading pair symbol. Defaults to BTC-USD if not provided.",
            "schema": {
              "type": "string",
              "default": "BTC-USD",
              "enum": ["BTC-USD", "ETH-USD", "XAU-USD", "SOL-USD", "BNB-USD"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with liquidity analysis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidityAnalysis"
                },
                "example": {
                  "symbol": "BTC-USD",
                  "score": 0.76,
                  "pools": [
                    { "level": 67850, "type": "demand", "strength": 0.85, "volume": "24.5 BTC" },
                    { "level": 69500, "type": "supply", "strength": 0.72, "volume": "18.2 BTC" }
                  ],
                  "sweeps": [
                    { "level": 67100, "direction": "bearish", "swept_at": "2025-01-14T14:30:00.000Z", "volume": "45.2 BTC" }
                  ],
                  "updated_at": "2025-01-15T14:30:00.000Z"
                }
              }
            },
            "headers": {
              "X-RateLimit-Limit": {
                "description": "Maximum number of requests allowed in the current period",
                "schema": { "type": "integer" }
              },
              "X-RateLimit-Remaining": {
                "description": "Number of requests remaining in the current period",
                "schema": { "type": "integer" }
              },
              "X-RateLimit-Reset": {
                "description": "Time when the rate limit resets (Unix timestamp)",
                "schema": { "type": "integer" }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Unauthorized",
                  "message": "API key required."
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/historical": {
      "get": {
        "operationId": "getHistoricalData",
        "summary": "Get historical market data",
        "description": "Returns historical market regime data for a given symbol over a specified time range. Useful for backtesting, charting, and trend analysis. Maximum lookback is 168 hours (7 days). Data points include regime classification, conviction, volatility, and liquidity score at each interval.",
        "tags": ["Historical Data"],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "description": "Trading pair symbol. Defaults to BTC-USD if not provided.",
            "schema": {
              "type": "string",
              "default": "BTC-USD",
              "enum": ["BTC-USD", "ETH-USD", "XAU-USD", "SOL-USD", "BNB-USD"]
            }
          },
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "description": "Number of hours of historical data to retrieve (1-168). Defaults to 24.",
            "schema": {
              "type": "integer",
              "default": 24,
              "minimum": 1,
              "maximum": 168
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "description": "Data interval granularity. Defaults to 1h.",
            "schema": {
              "type": "string",
              "default": "1h",
              "enum": ["5m", "15m", "30m", "1h", "4h", "1d"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with historical data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalResponse"
                },
                "example": {
                  "symbol": "BTC-USD",
                  "interval": "1h",
                  "data": [
                    {
                      "timestamp": "2025-01-15T13:00:00.000Z",
                      "regime": "trending_bullish",
                      "conviction": 0.78,
                      "volatility": "elevated",
                      "liquidity_score": 0.72
                    }
                  ],
                  "count": 25,
                  "timestamp": "2025-01-15T14:30:00.000Z"
                }
              }
            },
            "headers": {
              "X-RateLimit-Limit": {
                "description": "Maximum number of requests allowed in the current period",
                "schema": { "type": "integer" }
              },
              "X-RateLimit-Remaining": {
                "description": "Number of requests remaining in the current period",
                "schema": { "type": "integer" }
              },
              "X-RateLimit-Reset": {
                "description": "Time when the rate limit resets (Unix timestamp)",
                "schema": { "type": "integer" }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Unauthorized",
                  "message": "API key required."
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication. Pass your MarketMate API key as a Bearer token in the Authorization header."
      }
    },
    "schemas": {
...
      "LiquidityPool": {
        "type": "object",
        "required": ["level", "type", "strength", "volume"],
        "properties": {
          "level": {
            "type": "number",
            "description": "Price level of the liquidity pool",
            "example": 67850
          },
          "type": {
            "type": "string",
            "description": "Type of liquidity pool",
            "enum": ["demand", "supply"],
            "example": "demand"
          },
          "strength": {
            "type": "number",
            "description": "Pool strength score (0.0 - 1.0)",
            "minimum": 0,
            "maximum": 1,
            "example": 0.85
          },
          "volume": {
            "type": "string",
            "description": "Estimated volume at this level",
            "example": "24.5 BTC"
          }
        }
      },
      "LiquiditySweep": {
        "type": "object",
        "required": ["level", "direction", "swept_at", "volume"],
        "properties": {
          "level": {
            "type": "number",
            "description": "Price level that was swept",
            "example": 67100
          },
          "direction": {
            "type": "string",
            "description": "Direction of the sweep",
            "enum": ["bullish", "bearish"],
            "example": "bearish"
          },
          "swept_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp of when the sweep occurred",
            "example": "2025-01-14T14:30:00.000Z"
          },
          "volume": {
            "type": "string",
            "description": "Volume involved in the sweep",
            "example": "45.2 BTC"
          }
        }
      },
      "LiquidityAnalysis": {
        "type": "object",
        "required": ["symbol", "score", "pools", "sweeps", "updated_at"],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "Trading pair symbol",
            "example": "BTC-USD"
          },
          "score": {
            "type": "number",
            "description": "Overall liquidity score (0.0 - 1.0)",
            "minimum": 0,
            "maximum": 1,
            "example": 0.76
          },
          "pools": {
            "type": "array",
            "description": "Liquidity pools identified at key levels",
            "items": {
              "$ref": "#/components/schemas/LiquidityPool"
            }
          },
          "sweeps": {
            "type": "array",
            "description": "Recent liquidity sweeps",
            "items": {
              "$ref": "#/components/schemas/LiquiditySweep"
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp of when this data was last updated",
            "example": "2025-01-15T14:30:00.000Z"
          }
        }
      },
      "HistoricalDataPoint": {
        "type": "object",
        "required": ["timestamp", "regime", "conviction", "volatility", "liquidity_score"],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp of the data point",
            "example": "2025-01-15T13:00:00.000Z"
          },
          "regime": {
            "type": "string",
            "description": "Market regime at this point",
            "enum": ["trending_bullish", "trending_bearish", "ranging", "volatile"],
            "example": "trending_bullish"
          },
          "conviction": {
            "type": "number",
            "description": "Conviction score (0.0 - 1.0)",
            "minimum": 0,
            "maximum": 1,
            "example": 0.78
          },
          "volatility": {
            "type": "string",
            "description": "Volatility level",
            "enum": ["low", "normal", "elevated", "extreme"],
            "example": "elevated"
          },
          "liquidity_score": {
            "type": "number",
            "description": "Liquidity score (0.0 - 1.0)",
            "minimum": 0,
            "maximum": 1,
            "example": 0.72
          }
        }
      },
      "HistoricalResponse": {
        "type": "object",
        "required": ["symbol", "interval", "data", "count", "timestamp"],
        "properties": {
          "symbol": {
            "type": "string",
            "description": "Trading pair symbol",
            "example": "BTC-USD"
          },
          "interval": {
            "type": "string",
            "description": "Data interval used",
            "enum": ["5m", "15m", "30m", "1h", "4h", "1d"],
            "example": "1h"
          },
          "data": {
            "type": "array",
            "description": "Array of historical data points",
            "items": {
              "$ref": "#/components/schemas/HistoricalDataPoint"
            }
          },
          "count": {
            "type": "integer",
            "description": "Number of data points returned",
            "example": 25
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp of the response",
            "example": "2025-01-15T14:30:00.000Z"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type identifier",
            "example": "Unauthorized"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "API key required. Include Authorization: Bearer <your-api-key> header."
          }
        }
      },
      "RateLimitError": {
        "type": "object",
        "required": ["error", "message", "retry_after"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type identifier",
            "example": "RateLimitExceeded"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "You have exceeded your rate limit. Please retry after 60 seconds."
          },
          "retry_after": {
            "type": "integer",
            "description": "Seconds until the rate limit resets",
            "example": 60
          }
        }
      },
      "ServerError": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type identifier",
            "example": "InternalServerError"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "An unexpected error occurred. Please try again later."
          }
        }
      }
    }
  }
}